# openapi/handshake.v1.yaml
openapi: 3.1.0
info:
  title: Handshake API
  version: 1.0.0
  description: >
    Contract-first API surface for Handshake workflow, policy, metrics, and
    operator rails.
servers:
  - url: /
...
paths:
...
  /agents/{agentId}:
    get:
      tags:
        - operators
      summary: Agent detail
      operationId: getAgent
      parameters:
        - $ref: "#/components/parameters/AgentIdParam"
      responses:
        "200":
          description: Agent detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDetail"
        "404":
          $ref: "#/components/responses/ErrorResponse"
...
  /entities/{entityId}:
    get:
      tags:
        - operators
      summary: Entity detail
      operationId: getEntity
      parameters:
        - $ref: "#/components/parameters/EntityIdParam"
      responses:
        "200":
          description: Entity detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntityDetail"
        "404":
          $ref: "#/components/responses/ErrorResponse"
components:
  parameters:
...
    AgentIdParam:
      in: path
      name: agentId
      required: true
      schema:
        type: string
    EntityIdParam:
      in: path
      name: entityId
      required: true
      schema:
        type: string
  responses:
    ErrorResponse:
      description: Deterministic error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    DecisionEnum:
      type: string
      enum: [allow, deny, escalate]
...
    ErrorResponse:
      type: object
      required: [status, error, reasonCode, responseClass]
      properties:
        status: { type: string, enum: [error] }
        error: { type: string }
        reasonCode: { type: string }
        responseClass: { $ref: "#/components/schemas/ResponseClass" }
        message: { type: string }
      additionalProperties: false
...
    AgentDecision:
      type: object
      required: [requestId, timestamp, reasonCode, decision]
      properties:
        requestId: { type: string }
        timestamp: { type: number }
        reasonCode: { type: string }
        decision: { $ref: "#/components/schemas/DecisionEnum" }
    AgentDetail:
      type: object
      required: [agentId, riskTier, status, driftScore, recentRequests]
      properties:
        agentId: { type: string }
        riskTier: { type: string, enum: [low, medium, high, critical, unknown] }
        status: { type: string }
        driftScore: { type: number }
        recentRequests:
          type: array
          items: { $ref: "#/components/schemas/AgentDecision" }
...
    EntityInterface:
      type: object
      required: [interfaceId, kind, label, locator, verificationState, createdAt, updatedAt]
      properties:
        interfaceId: { type: string }
        kind: { type: string }
        label: { type: string }
        locator: { type: string }
        verificationState: { type: string }
        authMode: { type: string }
        createdAt: { type: number }
        updatedAt: { type: number }
    EntityRepresentation:
      type: object
      required: [representationId, agentId, principalId, scopes, issuedAt]
      properties:
        representationId: { type: string }
        agentId: { type: string }
        principalId: { type: string }
        scopes:
          type: array
          items: { type: string }
        interfaceIds:
          type: array
          items: { type: string }
        issuedAt: { type: number }
        expiresAt: { type: number }
        revokedAt: { type: number }
    EntityDetail:
      type: object
      required: [entityId, entityType, displayName, ownerPrincipalId, status, trustState, exposureScore, createdAt, updatedAt, interfaces, representations]
      properties:
        entityId: { type: string }
        entityType: { type: string }
        displayName: { type: string }
        legalName: { type: string }
        ownerPrincipalId: { type: string }
        status: { type: string }
        trustState: { type: string }
        exposureScore: { type: number }
        createdAt: { type: number }
        updatedAt: { type: number }
        interfaces:
          type: array
          items: { $ref: "#/components/schemas/EntityInterface" }
        representations:
          type: array
          items: { $ref: "#/components/schemas/EntityRepresentation" }
    ResponseClass:
      type: string
      enum: [ok, retryable, blocked, unknown]
  securitySchemes:
    IdentityEnvelopeHeader:
      type: apiKey
      in: header
      name: x-identity-envelope
      description: Canonical identity envelope (JSON) for protected routes.
    InternalTrustContextHeader:
      type: apiKey
      in: header
      name: x-internal-trust-context
      description: Signed internal trust token for privileged internal transitions.
security: []